fix: update code for renamed model classes
- Filter UnknownAgentTool in agent update (context.py)
- Update examples for renamed classes:
  - BatchRequest, UserMessage: import from models module
  - ClassifierTrainingParametersIn -> ClassifierTrainingParameters
  - ClassifierJobOut -> ClassifierFineTuningJob
  - Add type narrowing for ClassifierFineTuningJobDetails
