+++ examples/mistral/audio/transcription_async.py
@@ -2,7 +2,8 @@
 
 import os
 import asyncio
-from mistralai.client import Mistral, File
+from mistralai.client import Mistral
+from mistralai.client.models import File
 
 
 async def main():
